Add precedence support for the grammar
With no precedence taken into account, the evaluated expressions
give incorrect results. For example, the following expression
`1 + 1 * 2` would result in 4 instead of 3. The same goes with
boolean epxressions, `true or false and false` would result in
`false` instead of `true`.

This kind of behaviour is undesirable and should be corrected by
introducing precedence on operations.